+++ assets/js/gift-wrap-blocks.asset.php
@@ -1,0 +1,1 @@
+<?php return array('dependencies' => array('react-jsx-runtime', 'wc-blocks-checkout', 'wc-settings', 'wp-element', 'wp-plugins'), 'version' => 'b5b4e7e1b33721293cc3');